Abins 2D: move as much redundant code as possible to parent class
