+++ scripts/abins/abinsalgorithm.py
@@ -11,9 +11,10 @@ import os
 import re
 from typing import Tuple
 
-from mantid.api import mtd, FileAction, FileProperty, WorkspaceProperty
+import numpy as np
+from mantid.api import mtd, FileAction, FileProperty, WorkspaceGroup, WorkspaceProperty
 from mantid.kernel import Atom, Direction, StringListValidator, StringArrayProperty, logger
-from mantid.simpleapi import SaveAscii, Scale
+from mantid.simpleapi import CloneWorkspace, SaveAscii, Scale
 
 import abins
 from abins.constants import ALL_INSTRUMENTS, ALL_SAMPLE_FORMS, ATOM_PREFIX
@@ -222,18 +223,228 @@ class AbinsAlgorithm:
 
         return masses
 
-    def create_total_workspace(self, workspaces: list, force: bool = False) -> None:
-        """Sum together elemental totals to make an additional Total workspace
-
-        Nothing will be done if self._sum_contributions is False. This
-        can be overriden by setting force=True.
+    def create_workspaces(self, atoms_symbols=None, atom_numbers=None, *, s_data, max_quantum_order):
         """
+        Creates workspaces for all types of atoms. Creates both partial and total workspaces for given types of atoms.
+
+        :param atoms_symbols: atom types (i.e. element symbols) for which S should be created.
+        :type iterable of str:
+
+        :param atom_numbers:
+            indices of individual atoms for which S should be created. (One-based numbering; 1 <= I <= NUM_ATOMS)
+        :type iterable of int:
+
+        :param s_data: dynamical factor data
+        :type abins.SData
+
+        :param max_quantum_order: maximum quantum order to include
+        :type int:
+
+        :returns: workspaces for list of atoms types, S for the particular type of atom
+        """
+        from abins.constants import FLOAT_TYPE, MASS_EPS, ONLY_ONE_MASS
+
+        # Create appropriately-shaped arrays to be used in-place by _atom_type_s - avoid repeated slow instantiation
+        shape = [max_quantum_order]
+        shape.extend(list(s_data[0]["order_1"].shape))
+        s_atom_data = np.zeros(shape=tuple(shape), dtype=FLOAT_TYPE)
+        temp_s_atom_data = np.copy(s_atom_data)
+
+        num_atoms = len(s_data)
+        masses = self.get_masses_table(num_atoms)
+
+        result = []
+
+        if atoms_symbols is not None:
+            for symbol in atoms_symbols:
+                sub = (len(masses[symbol]) > ONLY_ONE_MASS
+                       or abs(Atom(symbol=symbol).mass - masses[symbol][0]) > MASS_EPS)
+                for m in masses[symbol]:
+                    result.extend(self._atom_type_s(num_atoms=num_atoms, mass=m, s_data=s_data,
+                                                    element_symbol=symbol, temp_s_atom_data=temp_s_atom_data,
+                                                    s_atom_data=s_atom_data, substitution=sub))
+        if atom_numbers is not None:
+            for atom_number in atom_numbers:
+                result.extend(self._atom_number_s(atom_number=atom_number, s_data=s_data,
+                                                  s_atom_data=s_atom_data))
+        return result
+
+    def _create_workspace(self, atom_name=None, s_points=None, optional_name="", protons_number=None,
+                          nucleons_number=None):
+        """
+        Creates workspace for the given frequencies and s_points with S data. After workspace is created it is rebined,
+        scaled by cross-section factor and optionally multiplied by the user defined scaling factor.
+
+
+        :param atom_name: symbol of atom for which workspace should be created
+        :param s_points: S(Q, omega)
+        :param optional_name: optional part of workspace name
+        :returns: workspace for the given frequency and S data
+        :param protons_number: number of protons in the given type fo atom
+        :param nucleons_number: number of nucleons in the given type of atom
+        """
+
+        ws_name = self._out_ws_name + "_" + atom_name + optional_name
+        self._fill_s_workspace(s_points=s_points, workspace=ws_name, protons_number=protons_number,
+                               nucleons_number=nucleons_number)
+        return ws_name
+
+    def _atom_number_s(self, atom_number=None, s_data=None, s_atom_data=None):
+        """
+        Helper function for calculating S for the given atomic index
+
+        :param atom_number: One-based index of atom in s_data e.g. 1 to select first element 'atom_1'
+        :type atom_number: int
+
+        :param s_data: Precalculated S for all atoms and quantum orders
+        :type s_data: abins.SData
+
+        :param s_atom_data: helper array to accumulate S (outer loop over atoms); does not transport
+            information but is used in-place to save on time instantiating large arrays. First dimension is quantum
+            order; following dimensions should match arrays in s_data.
+        :type s_atom_data: numpy.ndarray
+
+        :param
+
+        :returns: mantid workspaces of S for atom (total) and individual quantum orders
+        :returntype: list of Workspace2D
+        """
+        from abins.constants import ATOM_PREFIX, FUNDAMENTALS, S_LAST_INDEX
+
+        atom_workspaces = []
+        s_atom_data.fill(0.0)
+        output_atom_label = "%s_%d" % (ATOM_PREFIX, atom_number)
+        symbol = self._atoms_data[atom_number - 1]["symbol"]
+        z_number = Atom(symbol=symbol).z_number
+
+        for i, order in enumerate(range(FUNDAMENTALS, self._num_quantum_order_events + S_LAST_INDEX)):
+            s_atom_data[i] = s_data[atom_number - 1]["order_%s" % order]
+
+        total_s_atom_data = np.sum(s_atom_data, axis=0)
+
+        atom_workspaces = []
+        atom_workspaces.append(self._create_workspace(atom_name=output_atom_label,
+                                                      s_points=np.copy(total_s_atom_data),
+                                                      optional_name="_total", protons_number=z_number))
+        atom_workspaces.append(self._create_workspace(atom_name=output_atom_label,
+                                                      s_points=np.copy(s_atom_data),
+                                                      protons_number=z_number))
+        return atom_workspaces
+
+    def _atom_type_s(self, num_atoms=None, mass=None, s_data=None, element_symbol=None, temp_s_atom_data=None,
+                     s_atom_data=None, substitution=None):
+        """
+        Helper function for calculating S for the given type of atom
+
+        :param num_atoms: number of atoms in the system
+        :param s_data: Precalculated S for all atoms and quantum orders
+        :type s_data: abins.SData
+        :param element_symbol: label for the type of atom
+        :param temp_s_atom_data: helper array to accumulate S (inner loop over quantum order); does not transport
+            information but is used in-place to save on time instantiating large arrays.
+        :param s_atom_data: helper array to accumulate S (outer loop over atoms); does not transport
+            information but is used in-place to save on time instantiating large arrays.
+        :param substitution: True if isotope substitution and False otherwise
+        """
+        from abins.constants import FUNDAMENTALS, MASS_EPS, PYTHON_INDEX_SHIFT, S_LAST_INDEX
+
+        atom_workspaces = []
+        s_atom_data.fill(0.0)
+
+        element = Atom(symbol=element_symbol)
+
+        for atom_index in range(num_atoms):
+            if (self._atoms_data[atom_index]["symbol"] == element_symbol
+                    and abs(self._atoms_data[atom_index]["mass"] - mass) < MASS_EPS):
+
+                temp_s_atom_data.fill(0.0)
+
+                for order in range(FUNDAMENTALS, self._num_quantum_order_events + S_LAST_INDEX):
+                    order_indx = order - PYTHON_INDEX_SHIFT
+                    temp_s_order = s_data[atom_index]["order_%s" % order]
+                    temp_s_atom_data[order_indx] = temp_s_order
+
+                s_atom_data += temp_s_atom_data  # sum S over the atoms of the same type
+
+        total_s_atom_data = np.sum(s_atom_data, axis=0)
+
+        nucleons_number = int(round(mass))
+
+        if substitution:
+
+            atom_workspaces.append(self._create_workspace(atom_name=str(nucleons_number) + element_symbol,
+                                                          s_points=np.copy(total_s_atom_data),
+                                                          optional_name="_total", protons_number=element.z_number,
+                                                          nucleons_number=nucleons_number))
+            atom_workspaces.append(self._create_workspace(atom_name=str(nucleons_number) + element_symbol,
+                                                          s_points=np.copy(s_atom_data),
+                                                          protons_number=element.z_number,
+                                                          nucleons_number=nucleons_number))
+        else:
+
+            atom_workspaces.append(self._create_workspace(atom_name=element_symbol,
+                                                          s_points=np.copy(total_s_atom_data),
+                                                          optional_name="_total", protons_number=element.z_number))
+            atom_workspaces.append(self._create_workspace(atom_name=element_symbol,
+                                                          s_points=np.copy(s_atom_data),
+                                                          protons_number=element.z_number))
+
+        return atom_workspaces
+
+    def create_total_workspace(self, workspaces: list) -> None:
+        """Sum together elemental totals to make an additional Total workspace"""
         total_atom_workspaces = []
         for ws in workspaces:
             if "total" in ws:
                 total_atom_workspaces.append(ws)
         total_workspace = self._create_total_workspace(partial_workspaces=total_atom_workspaces)
         workspaces.insert(0, total_workspace)
+
+    def _create_total_workspace(self, partial_workspaces=None):
+        """
+        Sets workspace with total S.
+        :param partial_workspaces: list of workspaces which should be summed up to obtain total workspace
+        :returns: workspace with total S from partial_workspaces
+        """
+        from abins.constants import ONE_DIMENSIONAL_INSTRUMENTS, TWO_DIMENSIONAL_INSTRUMENTS
+        total_workspace = self._out_ws_name + "_total"
+
+        if isinstance(mtd[partial_workspaces[0]], WorkspaceGroup):
+            local_partial_workspaces = mtd[partial_workspaces[0]].names()
+        else:
+            local_partial_workspaces = partial_workspaces
+
+        if len(local_partial_workspaces) > 1:
+
+            # get frequencies
+            ws = mtd[local_partial_workspaces[0]]
+
+            # initialize S
+            if self._instrument.get_name() in ONE_DIMENSIONAL_INSTRUMENTS:
+                s_atoms = np.zeros_like(ws.dataY(0))
+
+            if self._instrument.get_name() in TWO_DIMENSIONAL_INSTRUMENTS:
+                n_q = abins.parameters.instruments[self._instrument.get_name()]['q_size']
+                n_energy_bins = ws.getDimension(1).getNBins()
+                s_atoms = np.zeros([n_q, n_energy_bins])
+
+            # collect all S
+            for partial_ws in local_partial_workspaces:
+                if self._instrument.get_name() in ONE_DIMENSIONAL_INSTRUMENTS:
+                    s_atoms += mtd[partial_ws].dataY(0)
+
+                elif self._instrument.get_name() in TWO_DIMENSIONAL_INSTRUMENTS:
+                    for i in range(n_energy_bins):
+                        s_atoms[:, i] += mtd[partial_ws].dataY(i)
+
+            # create workspace with S
+            self._fill_s_workspace(s_atoms, total_workspace)
+
+        # # Otherwise just repackage the workspace we have as the total
+        else:
+            CloneWorkspace(InputWorkspace=local_partial_workspaces[0], OutputWorkspace=total_workspace)
+
+        return total_workspace
 
     def write_workspaces_to_ascii(self) -> None:
         num_workspaces = mtd[self._out_ws_name].getNumberOfEntries()
@@ -267,6 +478,34 @@ class AbinsAlgorithm:
             cross_section = atom.neutron()["tot_scatt_xs"]
 
         return cross_section
+
+    @staticmethod
+    def set_workspace_units(wrk, layout='1D'):
+        """
+        Sets x and y units for a workspace.
+
+        :param layout: layout of data in Workspace2D.
+            - '1D' is a typical indirect spectrum, with energy transfer on Axis
+              0 (X), S on Axis 1 (Y)
+            - '2D' is a 2D S(q,omega) map with momentum transfer on Axis 0 (X),
+              S on Axis 1 and energy transfer on Axis 2
+        :type str:
+
+        :param wrk: workspace which units should be set
+        :type Workspace2D:
+        """
+
+        if layout == '1D':
+            mtd[wrk].getAxis(0).setUnit("DeltaE_inWavenumber")
+            mtd[wrk].setYUnitLabel("S / Arbitrary Units")
+            mtd[wrk].setYUnit("Arbitrary Units")
+        elif layout == '2D':
+            mtd[wrk].getAxis(0).setUnit("MomentumTransfer")
+            mtd[wrk].setYUnitLabel("S / Arbitrary Units")
+            mtd[wrk].setYUnit("Arbitrary Units")
+            mtd[wrk].getAxis(1).setUnit("DeltaE_inWavenumber")
+        else:
+            raise ValueError('Unknown data/units layout "{}"'.format(layout))
 
     @staticmethod
     def _validate_ab_initio_file_extension(*,
